add: Adjust npm run test with TypeScript
