+++ src/main.ts
@@ -1,12 +1,5 @@
-{
-  process.stdin.resume().setEncoding('utf8');
-  const lines: string[] = [];
-  require('readline')
-    .createInterface({ input: process.stdin })
-    .on('line', (input: string) => lines.push(input))
-    .on('close', main);
-
-  function main() {
-    // write code here
-  }
+function main(input: string[]) {
+  // write code here
 }
+
+export default main;
